reconstruct3d & colorizing - still fix #1

+++ src/finalcv-m10815822.hpp
@@ -16,10 +16,10 @@ public:
     std::vector<cv::Mat> RightExtrinsic;
     std::vector<cv::Mat> FMatrix;
 
-    std::vector<cv::Point2i> list_l;
-    std::vector<cv::Point2i> list_2d_color;
-    std::vector<cv::Point3i> list_3d_color;
-    std::vector<cv::Point3i> list_3d;
+    std::vector<cv::Point2f> list_l;
+    std::vector<cv::Point2f> list_2d_color;
+    std::vector<cv::Point3f> list_3d_color;
+    std::vector<cv::Point3f> list_3d;
 
     void readParams();
     void printinfo();
